Get rid of utils
