+++ ab2d-events-libs/src/main/java/gov/cms/ab2d/eventlibs/config/Ab2dEnvironment.java
@@ -1,4 +1,4 @@
-package gov.cms.ab2d.eventlibs.utils;
+package gov.cms.ab2d.eventlibs.config;
 
 import java.util.List;
 import java.util.stream.Stream;
